requirements-api.txt
langchain-community
pymongo
requests
python-dotenv
pypdf
flask
pytesseract
gunicorn
google-generativeai